+++ kitty/options/definition.py
@@ -1039,7 +1039,8 @@ MP3 or WAV on macOS (NSSound).
 opt('linux_bell_theme', '__custom', ctype='!bell_theme',
     long_text='''
 The XDG Sound Theme kitty will use to play the bell sound.
-Defaults to the custom theme name specified in the
+On Wayland, when the compositor supports it, it is asked to play the system default
+bell sound. Otherwise, defaults to the custom theme name specified in the
 :link:`XDG Sound theme specification <https://specifications.freedesktop.org/sound-theme-spec/latest/sound_lookup.html>,
 falling back to the default freedesktop theme if it does not exist.
 To change your sound theme desktop wide, create :file:`~/.local/share/sounds/__custom/index.theme` with the contents: